Legacy promotions: Conditionally run migrations
Since these promotions will be installed into all stores, we need to
make sure they don't do anything to the existing promotion system. For
new stores, they must do the things though.

+++ legacy_promotions/db/migrate/20190106184413_remove_code_from_spree_promotions.rb
@@ -9,22 +9,26 @@ class RemoveCodeFromSpreePromotions < ActiveRecord::Migration[5.1]
   end
 
   def up
-    promotions_with_code = Promotion.where.not(code: [nil, ''])
+    if column_exists?(:spree_promotions, :code)
+      promotions_with_code = Promotion.where.not(code: [nil, ''])
 
-    if promotions_with_code.any?
-      # You have some promotions with "code" field present! This is not good
-      # since we are going to remove that column.
-      #
-      self.class.promotions_with_code_handler.new(self, promotions_with_code).call
+      if promotions_with_code.any?
+        # You have some promotions with "code" field present! This is not good
+        # since we are going to remove that column.
+        #
+        self.class.promotions_with_code_handler.new(self, promotions_with_code).call
+      end
+
+      remove_index :spree_promotions, name: :index_spree_promotions_on_code
+      remove_column :spree_promotions, :code
     end
-
-    remove_index :spree_promotions, name: :index_spree_promotions_on_code
-    remove_column :spree_promotions, :code
   end
 
   def down
-    add_column :spree_promotions, :code, :string
-    add_index :spree_promotions, :code, name: :index_spree_promotions_on_code
+    unless column_exists?(:spree_promotions, :code)
+      add_column :spree_promotions, :code, :string
+      add_index :spree_promotions, :code, name: :index_spree_promotions_on_code
+    end
   end
 
   def self.promotions_with_code_handler
