+++ legacy_promotions/db/migrate/20230322085416_remove_match_policy_from_spree_promotion.rb
@@ -1,5 +1,7 @@
 class RemoveMatchPolicyFromSpreePromotion < ActiveRecord::Migration[5.2]
   def change
-    remove_column :spree_promotions, :match_policy, :string, default: "all"
+    if column_exists?(:spree_promotions, :match_policy)
+      remove_column :spree_promotions, :match_policy, :string, default: "all"
+    end
   end
 end
