+++ legacy_promotions/db/migrate/20230325132905_remove_unused_columns_from_promotion_rules.rb
@@ -1,7 +1,10 @@
 class RemoveUnusedColumnsFromPromotionRules < ActiveRecord::Migration[5.2]
   def change
-    remove_column :spree_promotion_rules, :code, :string
-
-    remove_column :spree_promotion_rules, :product_group_id, :integer
+    if column_exists?(:spree_promotion_rules, :code)
+      remove_column :spree_promotion_rules, :code, :string
+    end
+    if column_exists?(:spree_promotion_rules, :product_group_id)
+      remove_column :spree_promotion_rules, :product_group_id, :integer
+    end
   end
 end
